Remove PolicyView as it's not needed
