--- portals/publisher/src/main/webapp/source/src/app/components/GlobalPolicies/Policies/PolicyForm/Types.d.ts
@@ -1,40 +1,0 @@
-/*
-* Copyright (c) 2023, WSO2 LLC. (http://www.wso2.org) All Rights Reserved.
-* 
-* WSO2 LLC. licenses this file to you under the Apache License,
-* Version 2.0 (the "License"); you may not use this file except
-* in compliance with the License.
-* You may obtain a copy of the License at
-* 
-* http://www.apache.org/licenses/LICENSE-2.0
-* 
-* Unless required by applicable law or agreed to in writing,
-* software distributed under the License is distributed on an
-* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-* KIND, either express or implied.  See the License for the
-* specific language governing permissions and limitations
-* under the License.
-*/
-
-export type PolicyAttribute = {
-    id: string;
-    name: string;
-    displayName: string;
-    version: string;
-    description: string;
-    required: boolean;
-    type: string;
-    validationRegex: string;
-    defaultValue: any;
-    allowedValues: string[];
-};
-
-export type NewPolicyState = {
-    displayName: string | null;
-    version: string | null;
-    description: string;
-    applicableFlows: string[];
-    supportedGateways: string[];
-    supportedApiTypes: string[];
-    policyAttributes: any;
-};
